django-server/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
Django==3.0.4
djangorestframework==3.11.0
idna==2.9
kiwisolver==1.1.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.2
psycopg2==2.8.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
